chore: trim ponytail-audit dead weight
Drop unused shadcn, thin provider wrappers, dead learn/document
registry APIs, and collapse triplicate skill-flag resolvers.
