+++ apps/web/package.json
@@ -33,7 +33,6 @@
     "react-dom": "^19.1.0",
     "react-router-dom": "^7.18.2",
     "react-virtuoso": "^4.18.11",
-    "shadcn": "^4.8.0",
     "streamdown": "^2.5.0",
     "tailwind-merge": "^3.6.0",
     "tw-animate-css": "^1.4.0"
