+++ packages/core/package.json
@@ -48,7 +48,8 @@
     "./skills/archive": "./src/skills/archive.ts",
     "./skills/freshness": "./src/skills/freshness.ts",
     "./skills/bundled-names": "./src/skills/bundled-names.ts",
-    "./skills/post-turn-review": "./src/skills/post-turn-review.ts",
+    "./skills/post-turn-review": "./src/skills/profile-org-override.ts",
+    "./skills/profile-org-override": "./src/skills/profile-org-override.ts",
     "./skills/write": "./src/skills/write.ts",
     "./soul/org-memory": "./src/soul/org-memory.ts",
     "./tools/protected": "./src/tools/protected.ts",
