+++ packages/core/src/skills/index.ts
@@ -7,7 +7,6 @@ export * from "./bundled/install";
 export * from "./bundled-names";
 export * from "./compose";
 export * from "./consolidate";
-export * from "./curator-consolidate";
 export * from "./curator-schedule";
 export * from "./dedupe";
 export * from "./discover";
@@ -18,7 +17,6 @@ export * from "./load-tool";
 export * from "./match";
 export * from "./parse";
 export * from "./paths";
-export * from "./post-turn-review";
+export * from "./profile-org-override";
 export * from "./types";
 export * from "./write";
-export * from "./write-approval";
