+++ packages/core/src/skills/profile-org-override.ts
@@ -8,3 +8,45 @@ export function resolveProfileOrgBooleanOverride(
   }
   return orgValue === true;
 }
+
+export interface SkillWriteApprovalSources {
+  orgSkillsWriteApproval?: boolean | null;
+  profileSkillsWriteApproval?: boolean | null;
+}
+
+export function resolveSkillWriteApprovalRequired(
+  sources: SkillWriteApprovalSources
+): boolean {
+  return resolveProfileOrgBooleanOverride(
+    sources.profileSkillsWriteApproval,
+    sources.orgSkillsWriteApproval
+  );
+}
+
+export interface SkillPostTurnReviewSources {
+  orgSkillsPostTurnReview?: boolean | null;
+  profileSkillsPostTurnReview?: boolean | null;
+}
+
+export function resolveSkillPostTurnReviewEnabled(
+  sources: SkillPostTurnReviewSources
+): boolean {
+  return resolveProfileOrgBooleanOverride(
+    sources.profileSkillsPostTurnReview,
+    sources.orgSkillsPostTurnReview
+  );
+}
+
+export interface SkillCuratorConsolidateSources {
+  orgSkillsCuratorConsolidateEnabled?: boolean | null;
+  profileSkillsCuratorConsolidateEnabled?: boolean | null;
+}
+
+export function resolveSkillCuratorConsolidateEnabled(
+  sources: SkillCuratorConsolidateSources
+): boolean {
+  return resolveProfileOrgBooleanOverride(
+    sources.profileSkillsCuratorConsolidateEnabled,
+    sources.orgSkillsCuratorConsolidateEnabled
+  );
+}
