--- packages/core/src/skills/write-approval.ts
@@ -1,16 +1,0 @@
-import { resolveProfileOrgBooleanOverride } from "./profile-org-override";
-
-export interface SkillWriteApprovalSources {
-  orgSkillsWriteApproval?: boolean | null;
-  profileSkillsWriteApproval?: boolean | null;
-}
-
-/** Profile override wins when non-null; otherwise org default (false when unset). */
-export function resolveSkillWriteApprovalRequired(
-  sources: SkillWriteApprovalSources
-): boolean {
-  return resolveProfileOrgBooleanOverride(
-    sources.profileSkillsWriteApproval,
-    sources.orgSkillsWriteApproval
-  );
-}
